+++ config/splat.hd.dra.yaml
@@ -148,6 +148,7 @@ segments:
       - [0x419C0, .rodata, 7A4D0]
       - [0x41BD0, .rodata, 7E4BC]
       - [0x41CF0, .rodata, 843B0]
+      - [0x41DAC, .rodata, 8A0A4]
       - [0x41DCC, .rodata, 8D3E8]
       - [0x41E4C, .rodata, 90264]
       - [0x41F50, .rodata, 92F60]
@@ -179,6 +180,7 @@ segments:
       - [0x7A1C8, c, 7A4D0]
       - [0x7E168, c, 7E4BC]
       - [0x8403C, c, 843B0]
+      - [0x8A528, c, 8A0A4]
       - [0x8D050, c, 8D3E8]
       - [0x8FE50, c, 90264]
       - [0x91A98, c, 91EBC]
